% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globale_ovo.R
\name{aggregate_majoritaire_vote_ovo}
\alias{aggregate_majoritaire_vote_ovo}
\title{aggregate_majoritaire_vote_ovo}
\usage{
aggregate_majoritaire_vote_ovo(predictions_list)
}
\arguments{
\item{predictions_list:}{List of predictions}
}
\value{
vector of class aggregate
}
\description{
Function to aggregate one-versus-one predictions using majority voting.
}
