% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globale_ovo.R
\name{aggregate_predictions_Max_Voting}
\alias{aggregate_predictions_Max_Voting}
\title{aggregate_predictions}
\usage{
aggregate_predictions_Max_Voting(classes_list, score_list, y)
}
\arguments{
\item{classes_list:}{List of one versus all predictions}

\item{score_list:}{List of one versus all score}
}
\value{
vector of class aggregate
}
\description{
One versus all predictions aggregation function.
}
\details{
predictions aggregation function.
}
