% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globale_ovo.R
\name{aggregate_predictions_Max_Voting_ova}
\alias{aggregate_predictions_Max_Voting_ova}
\title{aggregate_predictions_Max_Voting_ova}
\usage{
aggregate_predictions_Max_Voting_ova(list_predictions, y)
}
\arguments{
\item{classes_list:}{List of one versus all predictions}

\item{score_list:}{List of one versus all score}
}
\value{
vector of class aggregate
}
\description{
Aggregation function of one-versus-all predictions using maximum score voting.
}
\details{
Aggregation function of one-versus-all predictions using maximum score voting.
}
