% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Analysis_Dataset_Complet_Sota_Predo}
\alias{Analysis_Dataset_Complet_Sota_Predo}
\title{Analysis_Dataset_Complet_Sota_Predo}
\format{
A \code{data.frame} with multiple rows and columns, including:
\describe{
\item{\code{dataset}}{Name of the evaluated dataset}
\item{\code{method}}{Method used (Predomics or SOTA)}
\item{\code{binarisation}}{Binarization strategy (e.g., "Terbeam")}
\item{\code{aggregation}}{Aggregation method (e.g., Tie, majority, max)}
\item{\code{constraint}}{Model constraint type (e.g., semi-constrained, etc)}
\item{\code{accuracy, recall, etc.}}{Performance metrics train/test}
}
}
\usage{
data(Analysis_Dataset_Complet_Sota_Predo)
}
\description{
Comprehensive dataset containing the results of all experiments conducted using Predomics and state-of-the-art (SOTA) methods.
}
\details{
This \code{data.frame} provides a structured summary of model performances across different configurations (OvA, OvO, constraints), datasets, and evaluation metrics. It is used for visualization (e.g., bar plots, forest plots) and statistical comparison (e.g., linear mixed models).
}
\author{
Edi Prifti, Eugeni Belda, Fabien Kambu
}
\keyword{benchmark,}
\keyword{model,}
\keyword{performance}
\keyword{predomics,}
\keyword{sota,}
