% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globale_ovo.R
\name{CalculateDistanceScores_}
\alias{CalculateDistanceScores_}
\title{CalculateDistanceScores_}
\usage{
CalculateDistanceScores_(mod)
}
\arguments{
\item{mod:}{object model}
}
\value{
list of max scores
}
\description{
This function calculates the maximum score  combinations..
}
\details{
This function calculates the maximum score combinations.
}
