% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terbeam.lib_mc.R
\name{countEachFeatureApperance_mc}
\alias{countEachFeatureApperance_mc}
\title{countEachFeatureApperance_mc}
\usage{
countEachFeatureApperance_mc(
  clf,
  allFeatures,
  pop,
  best,
  veryBest,
  approch = "ovo"
)
}
\arguments{
\item{clf}{A classifier object, containing model coefficients.}

\item{allFeatures}{A logical flag indicating whether all features are considered.}

\item{pop}{The entire population of models or features.}

\item{best}{The subset of the population representing the "best" models or features.}

\item{veryBest}{The subset of the population representing the "very best" models or features.}

\item{approch}{The approach for multiclass classification (default: "ovo").}
}
\value{
A list of data frames, where each data frame contains the number of appearances of
each feature in the "best" and "very best" populations, their normalized frequencies,
and their associated coefficients.
}
\description{
This function counts the frequency of feature appearances in a population,including their occurrences in the "best" and "very best" populations. It also evaluates the associated coefficients for each feature.
}
\details{
Count Each Feature's Appearance in a Multiclass Population
}
