% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globale_mc.R
\name{decompose_y_levels}
\alias{decompose_y_levels}
\title{Decompose Class Levels for Pairwise Combinations}
\usage{
decompose_y_levels(y)
}
\arguments{
\item{y}{A numeric or character vector representing the dependent variable with multiple classes.}
}
\value{
A list where each element contains the sorted samples for a specific pair of class levels.
}
\description{
This function generates pairwise combinations of class levels from a dependent variable \code{y} and sorts the samples within each pair.
}
