% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globale_ovo.R
\name{evaluateAdditionnalMetrics_ovo}
\alias{evaluateAdditionnalMetrics_ovo}
\title{evaluateAdditionnalMetrics_ovo}
\usage{
evaluateAdditionnalMetrics_ovo(mod, X, y, clf, mode = "train")
}
\arguments{
\item{mod:}{a model object to be evaluated}

\item{X:}{dataset to classify}

\item{y:}{variable to predict}

\item{clf:}{an object containing the different parameters of the classifier}

\item{mode:}{training or testing mode}
}
\value{
a list of models whose evaluation parameters are updated
}
\description{
This function computes prediction scores based on the confusion matrix such as accuracy, precision, recall and f-score
}
\details{
Compute other prediction scores such as precision, recall and f-score
}
