% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globale_ovo.R
\name{EvaluateAdditionnelGlobaleMetrics}
\alias{EvaluateAdditionnelGlobaleMetrics}
\title{EvaluateAdditionnelGlobaleMetrics}
\usage{
EvaluateAdditionnelGlobaleMetrics(predictions, actual_labels)
}
\arguments{
\item{actual_labels:}{current class (y)}

\item{aggregated}{predictions}
}
\value{
mod object
}
\description{
Overall additional evaluation metrics.
}
\details{
Overall additional evaluation metrics.
}
