% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globale_mc.R
\name{evaluateModel_aggregation}
\alias{evaluateModel_aggregation}
\title{evaluateModel_aggregation}
\usage{
evaluateModel_aggregation(mod, y)
}
\arguments{
\item{mod:}{The model object containing aggregated predictions and true labels.}

\item{y:}{The true class labels.}
}
\value{
The function returns the model object with evaluation metrics.
}
\description{
This function evaluates the aggregated model's performance.
}
