% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globale_ovo.R
\name{evaluateModels_aggregation}
\alias{evaluateModels_aggregation}
\title{evaluateModel_aggregation}
\usage{
evaluateModels_aggregation(
  mod,
  y,
  X,
  force.re.evaluation = TRUE,
  clf,
  approch = "ovo",
  aggregation_ = "votingAggregation"
)
}
\arguments{
\item{mod}{models objects.}

\item{y}{The true class labels.}

\item{X}{The feature matrix of the data to be predicted.}

\item{force.re.evaluation}{Boolean to force re-evaluation of the model even if it is already evaluated.}

\item{clf}{Object clf.}

\item{approach}{Type of approach to be used ("ovo" or "ova").}

\item{aggregation}{Type of aggregation method to be used ("votingAggregation", "weightedAggregation", "Predomics_aggregation_ova", "maximizationAggregation", "rankingAggregation").}
}
\value{
The function returns a list containing evaluated models with their respective metrics.
}
\description{
This function evaluates the aggregated models' performance.
}
