% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globale_ovo.R
\name{evaluatePopulation_aggregation}
\alias{evaluatePopulation_aggregation}
\title{evaluatePopulation_aggregation}
\usage{
evaluatePopulation_aggregation(
  pop,
  y,
  X,
  force.re.evaluation = TRUE,
  clf,
  aggregation = "votingAggregation"
)
}
\arguments{
\item{pop:}{A population of model objects.}

\item{y:}{The true class labels.}

\item{X:}{The feature matrix of the data to be predicted.}

\item{force.re.evaluation:}{Boolean to force re-evaluation of the model even if it is already evaluated.}

\item{clf:}{Object clf.}

\item{aggregation:}{Type of aggregation method to be used ("votingAggregation" or "weightedAggregation").}
}
\value{
The function returns a list containing evaluated models with their respective metrics.
}
\description{
This function evaluates the aggregated models' performance on a population level.
}
