% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globale_ovo.R
\name{evaluatePopulation_overall}
\alias{evaluatePopulation_overall}
\title{EvaluateAdditionnelGlobaleMetrics}
\usage{
evaluatePopulation_overall(pop, y, approch = "ova")
}
\arguments{
\item{pop}{: object pop}

\item{approch:}{approach}
}
\value{
overall pop object
}
\description{
Overall population evaluation.
}
\details{
Overall population evaluation
}
