% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globale_ovo.R
\name{evaluatesBestsModelsTest}
\alias{evaluatesBestsModelsTest}
\title{evaluatesBestsModelsTest}
\usage{
evaluatesBestsModelsTest(
  experiences,
  approch = "ovo",
  aggregation_ = "votingAggregation",
  clf = clf,
  y = y,
  X = X
)
}
\arguments{
\item{experiences}{experiences.}

\item{clf}{Object clf.}

\item{y}{y}

\item{X}{X}

\item{force.re.evaluation}{Boolean to force re-evaluation of the model even if it is already evaluated.}

\item{approach}{Type of approach to be used ("ovo" or "ova").}

\item{aggregation}{Type of aggregation method to be used ("votingAggregation", "weightedAggregation", "Predomics_aggregation_ova", "maximizationAggregation", "rankingAggregation").}
}
\value{
The function returns a list best modesls test evaluated
}
\description{
This function evaluates the models tests.
}
