% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globale_mc.R
\name{generate_combinations_with_factors}
\alias{generate_combinations_with_factors}
\title{Generate Class Combinations for Multiclass Classification}
\usage{
generate_combinations_with_factors(y, X, approch)
}
\arguments{
\item{y}{A numeric or character vector representing the dependent variable with multiple classes.}

\item{X}{A matrix or data frame where each column corresponds to the predictors associated with the values in \code{y}.}

\item{approch}{A character string specifying the approach to use: "ovo" for One-vs-One or "ova" for One-vs-All.}
}
\value{
A list containing:
}
\description{
This function generates combinations of classes and their corresponding data for binary classification approaches such as One-vs-One (OVO) and One-vs-All (OVA).
}
