% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terbeam.lib_mc.R
\name{generateAllCombinations_mc}
\alias{generateAllCombinations_mc}
\title{generateAllCombinations}
\usage{
generateAllCombinations_mc(
  X,
  y,
  clf,
  ind.features.to.keep,
  sparsity,
  allFeatures,
  approch = approch
)
}
\description{
Generate every possible combination of a list of features and evaluate them
}
