% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terbeam.lib_mc.R
\name{generateAllSingleFeatureModel_mc}
\alias{generateAllSingleFeatureModel_mc}
\title{generateAllSingleFeatureModel_mc}
\usage{
generateAllSingleFeatureModel_mc(
  X,
  y,
  clf,
  ind.sub = NULL,
  eval.all = FALSE,
  approch = "ovo"
)
}
\arguments{
\item{X}{A matrix of features, where columns correspond to samples.}

\item{y}{A vector of labels for the samples.}

\item{clf}{A classifier object with its parameters and coefficients.}

\item{ind.sub}{A list of subsets of indices for samples (default: NULL, which uses all samples).}

\item{eval.all}{A logical flag indicating whether to evaluate all models (default: FALSE).}

\item{approch}{A string indicating the approach for decomposition: "ovo" (default) or "ova".}
}
\value{
A list of models generated based on the specified approach.
}
\description{
This function generate All Single Feature Models for Multiclass Classificatio.
}
