% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terbeam.lib_mc.R
\name{getFeatures2Keep_mc}
\alias{getFeatures2Keep_mc}
\title{getFeatures2Keep_mc}
\usage{
getFeatures2Keep_mc(clf, featuresApperance, threshold = 0.01, approch = "ovo")
}
\arguments{
\item{clf}{A classifier object containing model parameters.}

\item{featuresApperance}{A data frame containing information about feature appearances,
including their frequency and coefficients.}

\item{threshold}{A numeric value specifying the minimum frequency required to retain a feature
(default: 0.01).}

\item{approch}{The approach for multiclass classification (default: "ovo").}
}
\value{
A list of features to keep for further processing.
}
\description{
This function determines which features should be retained based on their frequency of appearance in the "best" models and their associated coefficients. It applies a threshold to filter out less relevant features
}
\details{
Get Features to Keep in Multiclass Classification
}
