% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globale_ovo.R
\name{getSign_mc}
\alias{getSign_mc}
\title{getSign_mc}
\usage{
getSign_mc(X, y, clf = NULL, approch = "ovo", parallel.local = FALSE)
}
\arguments{
\item{X:}{the data matrix with variables in the rows and observations in the columns}

\item{y:}{the response vector}

\item{clf:}{the classifier parameter object}

\item{parallel.local:}{weather or not to run in //}
}
\value{
a list of +1 & -1 for each one-vs-all classification of our dataset
}
\description{
Evaluates the sign for a given feature this is the old getMgsVsTraitSignDiscr function, one versus one.
}
\details{
Evaluates the sign for a given feature this is the old getMgsVsTraitSignDiscr function
}
