% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globale_ovo.R
\name{listOfModelsToDenseCoefMatrix_mc}
\alias{listOfModelsToDenseCoefMatrix_mc}
\title{listOfModelsToDenseCoefMatrix}
\usage{
listOfModelsToDenseCoefMatrix_mc(
  clf,
  X,
  y,
  list.models,
  rm.empty = TRUE,
  order.row = TRUE
)
}
\arguments{
\item{clf:}{the classifier object}

\item{X:}{the dataset}

\item{y:}{the class vector}

\item{list.model:}{a list of model objects}

\item{rm.empty:}{remove null models in the list if any (default:TRUE)}

\item{order.row:}{order rows by occurence (default:TRUE)}
}
\value{
an data frame with model coefficients in rows
}
\description{
For each model in the list of models it will convert to dense format and convert to a data.frame
}
