% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globale_ovo.R
\name{listOfModelsToListOfDenseVec_mc}
\alias{listOfModelsToListOfDenseVec_mc}
\title{Builds a list of dense vector coefficients from a list of models}
\usage{
listOfModelsToListOfDenseVec_mc(clf, X, y, list.models)
}
\arguments{
\item{clf:}{classifier}

\item{X:}{dataset}

\item{y:}{labels}

\item{list.models:}{list of models}
}
\value{
a list of dense vectors of coefficient
}
\description{
Builds a list of dense vector coefficients from a list of models
}
