% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globale_ovo.R
\name{listOfSparseVecToListOfModels_mc}
\alias{listOfSparseVecToListOfModels_mc}
\title{listOfSparseVecToListOfModels_ovo}
\usage{
listOfSparseVecToListOfModels_mc(
  X,
  y,
  clf,
  v,
  lobj = NULL,
  eval.all = FALSE,
  approch = "ovo"
)
}
\arguments{
\item{X:}{dataset}

\item{y:}{labels}

\item{clf:}{classifier}

\item{v:}{list of vectors of coeffs. For example, v=list( c(0.0,1.0,0.0,-1.0) , c(1.0,1.0,0.0,0.0) , c(0.0,1.0,1.0,-1.0) )}

\item{lobj:}{a list of objects to add as elements in the model objects if not null (default:NULL)}

\item{eval.all:}{evaluate population (default:FALSE)}
}
\value{
an list of model object ovo and ova
}
\description{
Converts an list of "SparseVec" objects onto a list of predomics objects
}
\details{
listOfSparseVecToListOfModels
}
