% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globale_mc.R
\name{Majority_Voting_with_Tie_Breaking}
\alias{Majority_Voting_with_Tie_Breaking}
\title{Majority_Voting_with_Tie_Breaking}
\usage{
Majority_Voting_with_Tie_Breaking(mod)
}
\arguments{
\item{mod:}{The model object containing predictions and scores.}
}
\value{
The function returns the model object with aggregated predictions.
}
\description{
Function to aggregate one-versus-one predictions using majority voting tie breaking, incorporating the ability to break ties by selecting the class with the highest weighted score among the tied classes.
}
