% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globale_mc.R
\name{Majority_Voting_with_Tie_Breaking_ova}
\alias{Majority_Voting_with_Tie_Breaking_ova}
\title{Majority_Voting_with_Tie_Breaking_ova}
\usage{
Majority_Voting_with_Tie_Breaking_ova(mod, y)
}
\arguments{
\item{mod}{A list of submodels (OVA classifiers), each containing predictions and prediction scores.}

\item{y}{A vector of true labels or the complete set of possible classes.}
}
\value{
A list representing the aggregated model, including the final predicted classes.
}
\description{
Aggregates predictions from one-versus-all (OVA) classifiers using Majority_Voting_with_Tie_Breaking_ova.
In case of a tie between classes (equal number of votes), the class whose associated OVA model
provides the highest prediction score for the current sample is selected.
If no model predicts a class (all vote "ALL"), a class is selected randomly from the full class set.
}
