% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gobale_visuel_mc.R
\name{makeFeatureAnnot_mc}
\alias{makeFeatureAnnot_mc}
\title{Prints as text the detail on a given experiment along with summarized results (if computed)}
\usage{
makeFeatureAnnot_mc(pop, X, y, clf, approch = "ovo")
}
\arguments{
\item{pop:}{a population of models}

\item{X:}{the X dataset where to compute the abundance and prevalence}

\item{y:}{the target class}

\item{approch:}{approach ovo or ova}

\item{clf:}{an object containing the different parameters of the classifier}
}
\value{
a list with two data.frames one containing the coefficients per each model and the other a data.frame on the features
}
\description{
This function takes a population of models and creates a table with annotation on the features,
such as prevalence in the models and dataset as well as different statistics
}
