% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globale_mc.R
\name{maximization}
\alias{maximization}
\title{maximization}
\usage{
maximization(mod, y)
}
\arguments{
\item{mod:}{The model object containing predictions and scores.}

\item{y:}{True class.}
}
\value{
The function returns the model object with aggregated predictions.
}
\description{
Function to aggregate one-versus-all predictions using maximization Aggregation.
}
\details{
Aggregation function of one-versus-all predictions using maximum score.
}
