% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globale_mc.R
\name{maximization_accuracy}
\alias{maximization_accuracy}
\title{maximization_accuracy}
\usage{
maximization_accuracy(mod, y)
}
\arguments{
\item{mod:}{The model object containing predictions and scores.}

\item{y:}{True class.}
}
\value{
The function returns the model object with aggregated predictions.
}
\description{
Function to aggregate one-versus-all predictions using maximization aggregation, weighted by AUC and F1 score.
}
\details{
Aggregation function of one-versus-all predictions using maximum score, weighted by AUC and F1 score.
}
