% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globale_mc.R
\name{maximization_ovo}
\alias{maximization_ovo}
\title{maximization_ovo}
\usage{
maximization_ovo(mod)
}
\arguments{
\item{mod:}{The model object containing predictions and scores.}
}
\value{
The function returns the model object with aggregated predictions.
}
\description{
Function to aggregate one-versus-one predictions using maximization ovo.
}
\details{
This function performs maximization ovo aggregation of predictions.
}
