% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globale_ovo.R
\name{modelToDenseVec_mc}
\alias{modelToDenseVec_mc}
\title{Transform the model object onto dense format (long) one}
\usage{
modelToDenseVec_mc(natts, mod)
}
\arguments{
\item{natts:}{the number of attributes}

\item{mod:}{a predomics model object}
}
\value{
a dense (long) format model
}
\description{
Builds a model object based on model that is in the dense (long) format.
}
