% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globale_ovo.R
\name{normalize_scores}
\alias{normalize_scores}
\title{normalize_scores}
\usage{
normalize_scores(scores)
}
\arguments{
\item{mod:}{object model}
}
\value{
list of scores normalize
}
\description{
Function to normalize a vector between 0 and 1.
}
\details{
Function to normalize a vector between 0 and 1.
}
