% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gobale_visuel_ovo.R
\name{plotAUC_mc}
\alias{plotAUC_mc}
\title{Analyze the results from a list of classifiers}
\usage{
plotAUC_mc(scores, y, main = "", ci = TRUE, percent = TRUE, approch = "ova")
}
\arguments{
\item{scores:}{a list where each element is a vector of scores from a given model}

\item{y:}{the class to be predicted}

\item{main:}{title of the graph}

\item{ci:}{the point shape for the graph}

\item{percent:}{color for the graph}
}
\value{
a list of roc objects
}
\description{
Analyze the results from a list of classifiers.
}
