% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gobale_visuel_ovo.R
\name{plotFeatureModelCoeffs_mc}
\alias{plotFeatureModelCoeffs_mc}
\title{Plots the prevalence of a list of features in the whole dataset and per each class}
\usage{
plotFeatureModelCoeffs_mc(
  feat.model.coeffs,
  y,
  approch = "OVA",
  topdown = TRUE,
  col = c("deepskyblue1", "white", "firebrick1"),
  vertical.label = TRUE
)
}
\arguments{
\item{feat.model.coeffs:}{feature vs. model coeffient table}

\item{topdown:}{showing features from top-down or the other way around (default:TRUE)}

\item{col:}{colors to be used for the coeffients (default: -1 = deepskyblue1, 0 = white, 1 = firebrick1)}

\item{vertical.label:}{wether the x-axis labels should be vertical or not (default:TRUE)}
}
\value{
a list ggplot object
}
\description{
Plots the coefficients of subset of features in the models where they are found
}
