% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globale_ovo.R
\name{predict_ova}
\alias{predict_ova}
\title{predict_ova}
\usage{
predict_ova(mod, class_names)
}
\arguments{
\item{y:}{the response vector}

\item{mod:}{object model}
}
\value{
list of class predict
}
\description{
Function to predict the class for each "Class vs ALL" combination.
}
\details{
Function to predict the class for each "Class vs ALL" combination.
}
