% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globale_ovo.R
\name{predict_ovo}
\alias{predict_ovo}
\title{predict_ovo}
\usage{
predict_ovo(mod, y)
}
\arguments{
\item{y:}{the response vector}

\item{mod:}{object model}
}
\value{
list of class predict
}
\description{
Function to predict the class for each "Class1 vs Class2" combination.
}
\details{
Function to predict the class for each "Class 1 vs Class 2" combination.
}
