% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globale_ovo.R
\name{predictModel_ova}
\alias{predictModel_ova}
\title{predictModel_ova}
\usage{
predictModel_ova(mod, y, X, clf, force.re.evaluation = TRUE)
}
\arguments{
\item{y:}{The true class labels.}

\item{X:}{The feature matrix of the data to be predicted.}

\item{mod:}{The model object containing the OvA classification models.}

\item{clf:}{Object clf.}

\item{force.re.evaluation:}{Boolean to force re-evaluation of the model even if it is already evaluated.}
}
\value{
This function returns the list of one-versus-all predictions and the list of normalized scores.
}
\description{
This function predicts outcomes for a one-versus-all (OvA) classification model.
}
\details{
This function predicts outcomes for a one-versus-all (OvA) classification model.
}
