% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globale_mc.R
\name{predictModel_ovo}
\alias{predictModel_ovo}
\title{predictModel_ovo}
\usage{
predictModel_ovo(mod, y, X, clf, force.re.evaluation = TRUE)
}
\arguments{
\item{y:}{The true class labels}

\item{X:}{The feature matrix of the data to be predicted}

\item{mod:}{The model object containing the OvO classification models}

\item{clf:}{The clf object}

\item{force.re.evaluation:}{Boolean to force re-evaluation of the model even if it is already evaluated.}
}
\value{
This function returns mod with a list of one-versus-one predictions and the list of normalized scores.
}
\description{
This function predicts outcomes for a one-versus-one (OvO) classification model.
}
\details{
This function predicts outcomes for a one-versus-one (OvO) classification model.
}
