% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{predomics.inputs.ExperimentHubMulticlass}
\alias{predomics.inputs.ExperimentHubMulticlass}
\title{predomics.inputs.ExperimentHubMulticlass}
\format{
A named list of 16 datasets. Each item is a list with:
\describe{
\item{\code{X}}{Binary matrix of microbiome profiles}
\item{\code{y.df}}{Data frame with target labels (\code{study_condition})}
}
}
\usage{
data(predomics.inputs.ExperimentHubMulticlass)
}
\description{
Collection of real multiclass microbiome datasets from ExperimentHub, pre-processed for multiclass Predomics.
}
\details{
This object is a named list of 16 real-world datasets. Each sublist includes:
\describe{
\item{\code{X}}{A binary matrix (features × samples), indicating presence/absence}
\item{\code{y.df}}{A data.frame with the multiclass target variable \code{study_condition}}
}

Examples:
\itemize{
\item \strong{FengQ_2015} — 153 samples: 47 adenoma, 60 control, 46 CRC
\item \strong{KarlssonFH_2013} — 142 samples: 42 control, 48 IGT, 52 T2D
\item Other datasets: GhensiP_2019, HMP_2019_t2d, HanniganGD_2017, IaniroG_2022, LiJ_2014, LiJ_2017,
LiSS_2016, LoombaR_2017, MetaCardis_2020_a, ThomasAM_2018a, VatanenT_2016,
XieH_2016, YachidaS_2019, ZellerG_2014
}

All datasets are formatted for use in interpretable multiclass classification with Predomics.
}
\author{
Edi Prifti, Eugeni Belda, Fabien Kambu
}
\keyword{ExperimentHub}
\keyword{classification,}
\keyword{microbiome,}
\keyword{multiclass,}
