% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globale_mc.R
\name{Predomics_aggregation_ova}
\alias{Predomics_aggregation_ova}
\title{Predomics_aggregation_ova}
\usage{
Predomics_aggregation_ova(mod, y)
}
\arguments{
\item{mod:}{The model object containing predictions and scores.}

\item{y:}{true class.}
}
\value{
The function returns the model object with aggregated predictions.
}
\description{
Function to aggregate one-versus-all predictions using New approach.
}
\details{
Aggregation function of one-versus-all predictions using search and score max.
}
