% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globale_ovo.R
\name{Predomics_aggregation_ovo}
\alias{Predomics_aggregation_ovo}
\title{Predomics_aggregation_ovo}
\usage{
Predomics_aggregation_ovo(mod)
}
\arguments{
\item{mod:}{The model object containing predictions and scores.}
}
\value{
The function returns the model object with aggregated predictions.
}
\description{
Function to aggregate one-versus-one predictions using majority voting, incorporating the ability to break ties by selecting the class with the highest weighted score among the tied classes.
}
