% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gobale_visuel_mc.R
\name{printClassifier_mc}
\alias{printClassifier_mc}
\title{printClassifier_mc}
\usage{
printClassifier_mc(obj, indent = "\\t--- ")
}
\arguments{
\item{obj:}{a Classifier object}

\item{indent:}{a string (default:'tab---') that will precede each element of the object.}
}
\value{
NULL if the object is not a valid Classifier
}
\description{
This function prints a summary of a Classifier object.
}
\details{
Prints as text the detail on a given Classifier object
}
