% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gobale_visuel_mc.R
\name{printExperiment_mc}
\alias{printExperiment_mc}
\title{printExperiment_mc}
\usage{
printExperiment_mc(obj, indent = "\\t--- ")
}
\arguments{
\item{obj:}{an Experiment object}

\item{indent:}{a string (default:'tab---') that will precede each element of the object.}
}
\value{
NULL if the object is not a valid Experiment
}
\description{
This function prints a summary of an Experiment object.
}
\details{
Prints as text the detail on a given Experiment object
}
