% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gobale_visuel_mc.R
\name{printPopulation_mc}
\alias{printPopulation_mc}
\title{printPopulation_mc}
\usage{
printPopulation_mc(obj, method = "short", score = "fit_", indent = "")
}
\arguments{
\item{obj:}{a population of models to plot}

\item{method:}{if "digested" a short sumary (one line) will be printed, otherwise the method will contain the
specific way to print a model through the printModel() routine}

\item{score:}{which score to show in the fit (default:fit_)}

\item{indent:}{a string (default:'tab---') that will precede each element of the object.}
}
\description{
Prints a population of model multi class objects as text
}
\details{
Prints a population of model objects as text.
}
