% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gobale_visuel_ovo.R
\name{printy_mc}
\alias{printy_mc}
\title{Prints as text the detail on a given object from the predomics package.}
\usage{
printy_mc(obj)
}
\arguments{
\item{obj:}{an object from the predomics object}
}
\description{
This function will summarize any of the predomics package objects such as can be an Experiment,
a Model, a Population of models or a ModelCollection
}
