% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globale_ovo.R
\name{regenerate_clf}
\alias{regenerate_clf}
\title{regenerate_clf}
\usage{
regenerate_clf(clf, X, y, approch)
}
\arguments{
\item{clf}{Object representing the model to be regenerated. It should contain the fields required for updating, such as \code{feature.cor}.}

\item{X}{Matrix of input data (features) used for model training and testing.}

\item{y}{Vector of target labels corresponding to the rows of \code{X}.}

\item{approch}{Type of approach to be used for dataset decomposition. Accepts "ovo" (one-versus-one) or "ova" (one-versus-all).}
}
\value{
The function returns the updated \code{clf} object with new data and computed coefficients.
}
\description{
Regenerate the \code{clf} object with updated data and coefficients.
}
\details{
This function regenerates the \code{clf} object by initializing its data components, decomposing the dataset based on the specified approach, and computing the coefficients for the model.
}
