% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globale_mc.R
\name{sort_data}
\alias{sort_data}
\title{Sort Data by Dependent Variable
This function sorts a dependent variable \code{y} and its corresponding predictors \code{X} based on the ascending order of \code{y} values.}
\usage{
sort_data(y, X)
}
\arguments{
\item{y}{A numeric vector representing the dependent variable.}

\item{X}{A matrix or data frame where each column corresponds to the predictors associated with the values in \code{y}.}
}
\value{
A list containing:
}
\description{
Sort Data by Dependent Variable
This function sorts a dependent variable \code{y} and its corresponding predictors \code{X} based on the ascending order of \code{y} values.
}
