% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terga1_mc.R
\name{terga1_mc_fit}
\alias{terga1_mc_fit}
\title{terga1_mc_fit}
\usage{
terga1_mc_fit(
  X,
  y,
  clf,
  approch = "ova",
  aggregation_ = "Predomics_aggregation_ova",
  constraint_factor = "unconstrained"
)
}
\arguments{
\item{X}{A data frame or matrix containing the features of the dataset.
Each row corresponds to a sample, and each column corresponds to a feature.}

\item{y}{A vector containing the labels or target values. It must have the same length as the number of rows in \code{X}.}

\item{clf}{A classifier object that contains the model to be fitted. It must have a parameter list (\code{params})
that includes the classifier's configuration (e.g., language, objective, evaluation metric).}

\item{approch}{A character string indicating the approach to use for model evaluation.
Default is "ova" (one-vs-all), but other options may be available depending on the classifier.}

\item{aggregation_}{A character string specifying the aggregation method to be used. Default is "Predomics_aggregation_ova".}

\item{constrained}{A logical value indicating whether the classifier should be constrained in some way during fitting.
Default is FALSE.}
}
\value{
A fitted classifier model based on the genetic algorithm approach. The result can be used for predictions
and further analysis.
}
\description{
#' This function fits a classifier using a genetic algorithm-based approach.
It evolves populations of models and selects the best models according to a specified evaluation metric.
The classifier can be adapted to different types of language (binary, ternary, etc.),
and the function ensures that the evaluation metric matches the classifier's requirements.
}
