% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globale_mc.R
\name{updateModelIndex_mc}
\alias{updateModelIndex_mc}
\title{updateModelIndex}
\usage{
updateModelIndex_mc(obj, features = NULL)
}
\arguments{
\item{obj:}{the object is a model}

\item{features:}{the list of features which overrides the clf$data$features if this exists.}
}
\value{
the same object type as input, but updated
}
\description{
Update the index of a model objectn.
}
