% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globale_mc.R
\name{updateObjectIndex_mc}
\alias{updateObjectIndex_mc}
\title{updateObjectIndex}
\usage{
updateObjectIndex_mc(obj, features = NULL)
}
\arguments{
\item{obj:}{the object can be a model, population, or modelCollection}

\item{features:}{the list of features which overrides the clf$data$features if this exists.}
}
\value{
the same object type as input, but updated
}
\description{
Update the index of a model, population, or modelCollection.
}
