% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globale_mc.R
\name{voting}
\alias{voting}
\title{voting}
\usage{
voting(mod)
}
\arguments{
\item{mod:}{The model object containing predictions and scores.}
}
\value{
The function returns the model object with aggregated predictions.
}
\description{
Function to aggregate one-versus-one predictions using majority voting.
}
