% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globale_ovo.R
\name{votingAggregation}
\alias{votingAggregation}
\title{votingAggregation}
\usage{
votingAggregation(mod)
}
\arguments{
\item{mod:}{The model object containing predictions and scores.}
}
\value{
The function returns the model object with aggregated predictions.
}
\description{
Function to aggregate one-versus-one predictions using majority voting.
}
