% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globale_mc.R
\name{votingova}
\alias{votingova}
\title{votingova}
\usage{
votingova(mod, y)
}
\arguments{
\item{mod}{List of submodels containing predictions and metadata.}

\item{y}{Vector of class labels (true labels or full set of possible classes).}
}
\value{
A list representing the aggregated model object.
}
\description{
Function to aggregate one-versus-one or one-versus-all predictions using majority voting.
If there's a tie, a class is randomly selected. If no class is predicted, one is also randomly chosen among all possible.
}
