% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globale_mc.R
\name{weighted}
\alias{weighted}
\title{weighted}
\usage{
weighted(mod)
}
\arguments{
\item{mod:}{The model object containing predictions and scores.}
}
\value{
The function returns the model object with aggregated predictions.
}
\description{
Function to aggregate one-versus-one predictions using weighted voting.
}
\details{
This function performs weighted aggregation of predictions.
}
