% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globale_mc.R
\name{weighted_accuracy}
\alias{weighted_accuracy}
\title{weighted_accuracy}
\usage{
weighted_accuracy(mod)
}
\arguments{
\item{mod:}{The model object containing predictions and scores.}
}
\value{
The function returns the model object with aggregated predictions.
}
\description{
Function to aggregate one-versus-one predictions using weighted voting with Accuracy.
}
\details{
This function performs weighted aggregation of predictions, weighted by Accuracy.
}
