% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globale_mc.R
\name{weighted_f1_auc}
\alias{weighted_f1_auc}
\title{weighted_f1_auc}
\usage{
weighted_f1_auc(mod)
}
\arguments{
\item{mod:}{The model object containing predictions and scores.}
}
\value{
The function returns the model object with aggregated predictions.
}
\description{
Function to aggregate one-versus-one predictions using weighted voting with AUC and F1 score.
}
\details{
This function performs weighted aggregation of predictions, weighted by AUC and F1 score.
}
