% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mc.input}
\alias{mc.input}
\title{mc.input (frequencies)}
\description{
This dataset is a sample size of 894 consisting of 4 enterotypes
("Rum: 183", "Bact1: 299", "Bact2: 168", "Prev: 244") extracted from the metacardis database.
}
\author{
Edi Prifti, Egeuni Belda .
}
\keyword{METACARDIS}
\keyword{enterotypes,}
\keyword{microbiome,}
\keyword{species,}
