% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gobale_visuel_ovo.R
\name{printModel_mc}
\alias{printModel_mc}
\title{Prints a model object as text.}
\usage{
printModel_mc(mod, method = "short", score = "fit_")
}
\arguments{
\item{mod:}{a model to plot}

\item{method:}{an object containing the different parameters of the classifier}

\item{score:}{which score to show in the fit (default:fit_)}
}
\description{
Prints a model object as text
}
