% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predomics_OVO.R
\name{runClassifier_mc}
\alias{runClassifier_mc}
\title{runClassifier_ovo}
\usage{
runClassifier_mc(
  X,
  y,
  clf,
  x_test = NULL,
  y_test = NULL,
  approch = "ovo",
  aggregation_ = "votingAggregation"
)
}
\arguments{
\item{X:}{The dataset to classify}

\item{y:}{The variable to predict}

\item{clf:}{The classifier object containing the different settings of the classifier.}

\item{x_test:}{if not NULL (default) this dataset will be used to evaluate the models in a subset for the feature importance}

\item{y_test:}{if not NULL (default) this dataset will be used to evaluate the models in a subset for the feature importance}
}
\value{
the classifier along with the classification results as a sub-element
}
\description{
This function runs a classifier in a given dataset
}
\details{
Runs the learning on a dataset
}
